Lets make this plugin open-source
**Why did this plugin become open source?**
Well, I suffer a lot from the plugin to provide a better experience, but I can't do many things, so if anyone knows, it can help with PR opening.

The other thing is that it makes no sense to hide the plugin's source code, because it's not a big deal, as there is no premium thing you can't do without the Bukkit API.

The fact is, this plugin has no contract, so code modification/improve/conversion is possible, but no theft or loss.

This source is the current development version, that worked since the last update.