Fixed when the configuration keys and comments re-ordered and comments was duplicated upon reload
Upstream (Spigot) has introduced a way to process and parse comments when saving our configuration. With this change the saveToString method broken making it re-order the config keys and it includes comments which is also confused the loading process of TL. Now the parsing comments using ConfigurationOption is disabled, we uses our own implementation to include comments.

Fixes #307

Also, made a small micro optimization for retrieving player's ping and server tps, now it is only checked once if the method is exist.