Seems priority for groups is unnecessary (don't really know what will be the solution)
When the groups are loaded, priority comes into play and sorts the groups and creates the groups in the order the user has set. This worked for me with 2 players. I just removed the priority from the group name, maybe it would work without it, but since it works for some and not for others, I don't know. #339

Updated gradle wrapper to 7.5