Revert the proper sorting for groups
We've made some tests which includes that the sorting with team names works with this implementation. So the question still remains, why does it work for some and not for others? I still don't know the answer to that. Maybe it depends on the behaviour of the clients, I don't know.