Added option to measure server performance if the server is under heavy load
With this, we can achieve a small server performance improvement (not much), if the server is currently under heavy load. TabList will cancels all currently running processes, if the value is set, let's assume 15.0, when this value is reached by TPS, TabList will start cancelling all its processes. Once these have happened, TabList will not restart its processes if the TPS has reached an optimal value, so it must be done manually, by reconnecting to the server or by reloading the plugin.

We need such an option because TabList's processes run very quickly in the background and use a lot of resources, especially the slower running processes.