Warn the user if the CPU cores is less than 3
We are getting more and more reports of "server lags, tps drop etc" which is because the CPU threads (number of cores) are 2 or less so the server can't allocate/create the threads so that the tablist and server tps properly run without drop. Therefore, from now on, the user will receive a warning that their CPU has few cores (so it can only handle few threads).