Started refactoring NMS packets
Now we're accessing the NMS directly to use objects instead of depending on reflections. This means almost all the tablist's functionality will be reworked and legacy versions will be separated from the recent version.

Currently fake players and tablist objectives don't work, these 2 remained only.

Overall I decided to get rid of depending on reflection to make the process a bit faster. NMS classes from reflection will not be loaded and cached anymore on 1.19.3.

Also cleaned up the mess in gradle build files and switched to kotlin DSL for better readability and functionality